Fix wrong order of arguments to q.setRPY in commit 82964f6. The API shows the order is: void    setRPY (const tfScalar &roll, const tfScalar &pitch, const tfScalar &yaw)
Line 14 should read:

q.setRPY(0, 0, msg->theta);

This fault breaks this tutorial:

http://wiki.ros.org/tf/Tutorials/Writing%20a%20tf%20broadcaster%20%28C%2B%2B%29

In a way that doesn't show up until the subsequent tutorial, where the
follower doesn't follow. The fix to line 14 proposed above fixes the
subsequent tutorial.

The problem is discussed in this thread:

http://answers.ros.org/question/45884/turtle-tf-tutorial-fails-to-work/?answer=47845#post-id-47845

and an incorrect fix was applied to the wiki. The note in the wiki under
the code, which says the code should read:

transform.setRotation( tf::Quaternion(0, 0, msg->theta) );

is wrong - it should say transform.setRPY.