Remove the dependency on tf_transformations.
That way we don't have to install a non-standard library
(tf_transformations) that depends on a non-packaged pip
package (transforms3d) for the tutorials to work.

Note well that the function is copied twice, once in
static_turtle_tf2_broadcaster.py and once in turtle_tf2_broadcaster.py.
This is by design, as these files are used as standalone
examples in the ROS 2 tutorials.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>